backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI framework
jax
jaxlib
flax

# Data processing
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Blockchain
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.0
pyyaml==6.0.1

# Security
cryptography>=41.0.0
pyjwt>=2.8.0
bcrypt>=4.0.0

# Utils
pydantic>=2.5.0
python-dotenv==1.0.0

# Logging
loguru>=0.7.0
opencv-python-headless==4.10.0.84

# Hardware
pyserial==3.5

# Async
celery>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
requests
typing_extensions
